+++ src/modules/react-redux-hash-router4/RouterConstants.js
@@ -1,5 +1,4 @@
 export default{
-
-    NAME: 'ufp-react-redux-router4'
-
+    NAME: 'ufp-react-redux-router4',
+    REDUCER_NAME: 'router'
 }